Added testdoc to test format of commit message
Skipped a line, and how I'm going to write a lot more,
and save by simply writing the file in vim. I don't know
if this will work but I'm going to pus it up to github
and see if it worksssssssssss. I guess I can only just
hope that it does!